fix(saves): load save data now has quotes, safe plugins don't needlessly terminate

+++ src/omegga/plugin/plugin_node_safe.js
@@ -333,7 +333,8 @@ class NodeVmPlugin extends Plugin {
       this.#outInterface.removeAllListeners('line');
       this.#errInterface.removeAllListeners('line');
       try {
-        this.#worker.terminate();
+        if (this.#worker)
+          this.#worker.terminate();
       } catch (err) {
         Omegga.error('!>'.red, 'Error terminating worker for', this.getName().brightRed.underline, err);
       }
